requirements.txt
python-telegram-bot>=20.0
mysql-connector-python>=8.0.0
requests>=2.28.0
PyPDF2>=3.0.0
asyncio
aiohttp>=3.8.0
python-dotenv>=1.0.0
